+++ jsonpath/function_extensions/__init__.py
@@ -1,0 +1,3 @@
+from .length import length
+
+__all__ = ("length",)