+++ jsonpath/function_extensions/length.py
@@ -1,0 +1,9 @@
+from collections.abc import Sized
+from typing import Union
+
+
+def length(obj: Sized) -> Union[int, None]:
+    try:
+        return len(obj)
+    except TypeError:
+        return None